+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.1.66
 
 - fix remove games section on Netflix on chrome
+- fixed bug on prime chrome home page not filter paid content
 
 ## 1.1.65